S-13d extract agent modals to web/modules/agent-modals.js
Move terminal modal (xterm.js), conversation (readable transcript) modal and
agent reauth login flow from app.js into a dedicated ES module
web/modules/agent-modals.js. The module exports openTerminalModal,
openConversationModal, handleAgentLogin and initAgentModals; initAgentModals
receives openModal/closeModal/loadAgents via DI and wires all event listeners.
Sets window.handleAgentLogin so agents.js event listeners can reach it across
the ES module boundary. Add agent-modals-ui-contract.test.ts to guard wiring.
