+++ web/app.js
@@ -20,6 +20,7 @@ import { initTokenUsage, loadTokenUsage } from './modules/token-usage.js'
 import { startActivityPoll, stopActivityPoll, loadActivity, loadOverview, initActivity } from './modules/overview.js'
 import { loadFederationPage, initFederation } from './modules/federation.js'
 import { loadIdeasPage, initIdeas } from './modules/ideas.js'
+import { openTerminalModal, openConversationModal, initAgentModals } from './modules/agent-modals.js'
 
 // avatarBust() is imported from ./modules/agents.js (avatar epoch owned there).
 
@@ -321,6 +322,7 @@ initTokenUsage()
 initActivity({ openTerminalModal })
 initFederation({ openModal, closeModal })
 initIdeas({ openModal, closeModal })
+initAgentModals({ openModal, closeModal, loadAgents })
 
 // === Helpers ===
 function escapeHtml(str) {
@@ -2186,335 +2188,6 @@ async function _resolveApproval(id, decision) {
 
   checkStatus()
 })()
-
-// === Agent reauth login flow ===
-async function handleAgentLogin(agentName, btn) {
-  const phase = btn.dataset.phase || 'start'
-  btn.disabled = true
-  const origText = btn.textContent
-  btn.textContent = phase === 'start' ? t('agents.auth.btn_starting') : t('agents.auth.btn_confirming')
-  try {
-    const res = await fetch(`/api/agents/${encodeURIComponent(agentName)}/login`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ phase }),
-    })
-    if (!res.ok) { const err = await res.json().catch(() => ({})); throw new Error(err.error || 'HTTP ' + res.status) }
-    if (phase === 'start') {
-      btn.dataset.phase = 'confirm'
-      btn.textContent = t('agents.auth.btn_confirm')
-      btn.disabled = false
-      showToast(t('agents.auth.toast_started'))
-    } else {
-      btn.textContent = t('agents.auth.btn_logged_in')
-      showToast(t('agents.auth.toast_success'))
-      setTimeout(() => loadAgents(), 1500)
-    }
-  } catch (e) {
-    showToast('Hiba: ' + (e.message || e))
-    btn.textContent = origText
-    btn.dataset.phase = 'start'
-    btn.disabled = false
-  }
-}
-
-// === Agent terminal modal (xterm.js) ===
-let terminalInstance = null
-let terminalSSE = null
-let terminalFit = null
-// Master input gate (mirrors the server-side terminal-input toggle). Keystrokes
-// are dropped locally when OFF so we never spam the audit log with 403s; the
-// server enforces the same gate independently (fail-closed). Owner flips it via
-// the checkbox in the modal header (POST /api/terminal-input).
-let terminalInputEnabled = false
-
-function syncTerminalInputToggleUI() {
-  const cb = document.getElementById('terminalInputToggle')
-  const label = document.getElementById('terminalInputToggleLabel')
-  if (cb) cb.checked = terminalInputEnabled
-  if (label) {
-    label.textContent = terminalInputEnabled ? 'Input on' : 'Input off'
-    label.style.color = terminalInputEnabled ? '#8fbf6f' : '#b8b2a6'
-  }
-}
-
-function openTerminalModal(agentName) {
-  const overlay = document.getElementById('terminalOverlay')
-  const container = document.getElementById('terminalContainer')
-  const title = document.getElementById('terminalModalTitle')
-  if (!overlay || !container) return
-
-  title.textContent = agentName + ' - Terminal'
-
-  // Read the current server-side gate so the modal reflects reality on open.
-  fetch('/api/terminal-input')
-    .then(r => r.ok ? r.json() : { enabled: false })
-    .then(d => { terminalInputEnabled = d.enabled === true; syncTerminalInputToggleUI() })
-    .catch(() => { terminalInputEnabled = false; syncTerminalInputToggleUI() })
-
-  // Cleanup previous
-  if (terminalSSE) { terminalSSE.close(); terminalSSE = null }
-  if (terminalInstance) { terminalInstance.dispose(); terminalInstance = null }
-  container.innerHTML = ''
-
-  // Init xterm — fontSize 12 + wider modal fits ~140 chars of tmux output
-  const term = new window.Terminal({
-    theme: { background: '#1a1a1a', foreground: '#e8e4da' },
-    fontFamily: 'JetBrains Mono, Menlo, monospace',
-    fontSize: 12,
-    cursorBlink: false,
-    disableStdin: false,
-    scrollback: 4000,
-    convertEol: true,
-    allowProposedApi: true,
-  })
-  const fitAddon = new window.FitAddon.FitAddon()
-  term.loadAddon(fitAddon)
-  term.open(container)
-  fitAddon.fit()
-  terminalInstance = term
-  terminalFit = fitAddon
-
-  openModal(overlay)
-  setTimeout(() => term.focus(), 50)
-
-  // SSE pane stream.
-  // The pane snapshot now includes scrollback history (server uses
-  // `capture-pane -S -2000`), so the user can scroll back. To keep scrolling
-  // stable we (a) only repaint when the snapshot actually changed, and (b) only
-  // repaint while the viewport is at the bottom — if the user has scrolled up we
-  // freeze their view and resume painting when they return to the bottom (the
-  // onScroll handler below). The repaint clears the scrollback (CSI 3 J) before
-  // rewriting the full snapshot so frames don't accumulate duplicate history.
-  let latestPane = null
-  let paintedPane = null
-  const isAtBottom = () => {
-    const buf = term.buffer.active
-    return buf.viewportY >= buf.baseY
-  }
-  const repaint = () => {
-    if (latestPane === null || latestPane === paintedPane) return
-    if (!isAtBottom()) return // user scrolled up — keep their view put
-    paintedPane = latestPane
-    term.write('\x1b[3J\x1b[2J\x1b[H' + latestPane)
-  }
-  // EventSource cannot set an Authorization header. In token mode we pass the
-  // token via ?token=; in password-login (session-cookie) mode there is no
-  // token, so we open a plain URL and the browser attaches the mv_session
-  // cookie automatically -- the gate's cookie branch covers the SSE path.
-  const token = localStorage.getItem('marveen-dashboard-token') || ''
-  const streamBase = `/api/agents/${encodeURIComponent(agentName)}/pane/stream`
-  const sse = new EventSource(token ? `${streamBase}?token=${encodeURIComponent(token)}` : streamBase)
-  sse.onmessage = (e) => {
-    try {
-      const msg = JSON.parse(e.data)
-      if (msg.pane !== undefined) {
-        latestPane = msg.pane.replace(/\x1b]8;[^\x1b]*\x1b\\/g, '')
-        repaint()
-      }
-    } catch {}
-  }
-  sse.onerror = () => term.write(`\r\n${t('terminal.stream_error')}\r\n`)
-  terminalSSE = sse
-  // When the user scrolls back down to the bottom, resume live repainting.
-  term.onScroll(() => { if (isAtBottom()) repaint() })
-
-  // Single onData handler — maps escape sequences to {special}, plain chars to {keys}
-  // Using onData only (no onKey) avoids double-firing on arrow/Enter keys.
-  // PageUp/PageDown are intentionally NOT forwarded: they scroll the xterm
-  // scrollback locally (history viewing) instead of going to the agent.
-  const ESC_TO_SPECIAL = {
-    '\r': 'Enter', '\x1b': 'Escape',
-    '\x1b[A': 'Up', '\x1b[B': 'Down', '\x1b[C': 'Right', '\x1b[D': 'Left',
-    '\x7f': 'BSpace', '\t': 'Tab', '\x1b[Z': 'S-Tab',
-    '\x03': 'C-c', '\x04': 'C-d', '\x15': 'C-u', '\x0c': 'C-l',
-  }
-  term.onData(data => {
-    if (data === '\x1b[5~') { term.scrollPages(-1); return } // PageUp -> scroll history up
-    if (data === '\x1b[6~') { term.scrollPages(1); return }  // PageDown -> scroll history down
-    if (!terminalInputEnabled) {
-      // Read-only mode: input gate is OFF. Drop the keystroke locally (server
-      // would 403 it anyway) and nudge the user to the toggle.
-      showToast('Terminal input is off. Enable it with the header toggle first.')
-      return
-    }
-    const special = ESC_TO_SPECIAL[data]
-    const body = special ? { special } : { keys: data }
-    fetch(`/api/agents/${encodeURIComponent(agentName)}/keys`, {
-      method: 'POST', headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(body),
-    }).catch(() => {})
-  })
-
-  // Resize fit on modal resize — observe the modal wrapper (not the xterm container
-  // itself) to avoid a ResizeObserver->fit->resize->ResizeObserver infinite loop
-  let fitTimer = null
-  const ro = new ResizeObserver(() => {
-    clearTimeout(fitTimer)
-    fitTimer = setTimeout(() => { try { fitAddon.fit() } catch {} }, 50)
-  })
-  const modalEl = container.closest('.terminal-modal') || container.parentElement
-  if (modalEl) ro.observe(modalEl)
-}
-
-document.getElementById('terminalClose')?.addEventListener('click', () => {
-  const overlay = document.getElementById('terminalOverlay')
-  if (overlay) closeModal(overlay)
-  if (terminalSSE) { terminalSSE.close(); terminalSSE = null }
-  if (terminalInstance) { terminalInstance.dispose(); terminalInstance = null }
-})
-
-// Owner flips the master terminal-input gate. Optimistically reflect the desired
-// state, POST it, then reconcile with the server's authoritative response.
-document.getElementById('terminalInputToggle')?.addEventListener('change', (e) => {
-  const desired = e.target.checked === true
-  fetch('/api/terminal-input', {
-    method: 'POST', headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify({ enabled: desired }),
-  })
-    .then(r => r.ok ? r.json() : Promise.reject(new Error('HTTP ' + r.status)))
-    .then(d => {
-      terminalInputEnabled = d.enabled === true
-      syncTerminalInputToggleUI()
-      showToast(terminalInputEnabled ? 'Terminal input enabled (audit-logged)' : 'Terminal input disabled')
-    })
-    .catch(() => {
-      terminalInputEnabled = false
-      syncTerminalInputToggleUI()
-      showToast('Could not change terminal input state')
-    })
-})
-
-// === Agent conversation (readable transcript) modal ===
-// Renders the agent's Claude Code transcript as a chat-style timeline: inbound
-// Telegram messages, the agent's replies, and (optionally) its notes/actions.
-// Solves what the raw terminal can't: a readable, searchable review of what
-// actually happened -- also the support view for customer-hosted Marveens.
-const CONVERSATION_PAGE_SIZE = 400
-let conversationEntries = []
-let conversationAgentName = null
-let conversationHasOlder = false
-let conversationLoadingOlder = false
-
-async function openConversationModal(agentName, displayName) {
-  const overlay = document.getElementById('conversationOverlay')
-  const container = document.getElementById('conversationContainer')
-  const title = document.getElementById('conversationModalTitle')
-  if (!overlay || !container) return
-  conversationAgentName = agentName
-  title.textContent = t('conversation.title', { name: displayName || agentName })
-  container.innerHTML = `<div class="conversation-empty">${t('conversation.loading')}</div>`
-  openModal(overlay)
-  await loadConversation()
-}
-
-// Latest page (offset=0); resets the loaded window.
-async function loadConversation() {
-  const container = document.getElementById('conversationContainer')
-  const token = localStorage.getItem('marveen-dashboard-token') || ''
-  try {
-    const r = await fetch(`/api/agents/${encodeURIComponent(conversationAgentName)}/conversation?limit=${CONVERSATION_PAGE_SIZE}&offset=0`, {
-      headers: { 'Authorization': 'Bearer ' + token },
-    })
-    const d = await r.json()
-    conversationEntries = Array.isArray(d.entries) ? d.entries : []
-    conversationHasOlder = !!d.hasOlder
-    renderConversation()
-  } catch {
-    if (container) container.innerHTML = `<div class="conversation-empty">${t('conversation.error')}</div>`
-  }
-}
-
-// Page further back: fetch the window of entries immediately before the oldest
-// loaded one and PREPEND it, keeping the scroll position so the view does not
-// jump. Lets the operator read history beyond the on-screen window (and beyond
-// the old fixed cap).
-async function loadOlderConversation() {
-  if (conversationLoadingOlder || !conversationHasOlder) return
-  conversationLoadingOlder = true
-  const btn = document.getElementById('conversationLoadOlder')
-  if (btn) { btn.disabled = true; btn.textContent = t('conversation.loading') }
-  const token = localStorage.getItem('marveen-dashboard-token') || ''
-  try {
-    const offset = conversationEntries.length
-    const r = await fetch(`/api/agents/${encodeURIComponent(conversationAgentName)}/conversation?limit=${CONVERSATION_PAGE_SIZE}&offset=${offset}`, {
-      headers: { 'Authorization': 'Bearer ' + token },
-    })
-    const d = await r.json()
-    const older = Array.isArray(d.entries) ? d.entries : []
-    conversationHasOlder = !!d.hasOlder
-    if (older.length) {
-      conversationEntries = older.concat(conversationEntries)
-      renderConversation({ preserveScroll: true })
-    } else {
-      renderConversation()
-    }
-  } catch {
-    if (btn) { btn.disabled = false; btn.textContent = t('conversation.load_more') }
-  } finally {
-    conversationLoadingOlder = false
-  }
-}
-
-function fmtConvTs(ts) {
-  if (!ts) return ''
-  try {
-    return new Date(ts).toLocaleString('hu-HU', { month: '2-digit', day: '2-digit', hour: '2-digit', minute: '2-digit' })
-  } catch { return '' }
-}
-
-function renderConversation(opts = {}) {
-  const container = document.getElementById('conversationContainer')
-  if (!container) return
-  const prevH = container.scrollHeight
-  const prevTop = container.scrollTop
-  const q = (document.getElementById('conversationSearch')?.value || '').toLowerCase().trim()
-  const showActions = document.getElementById('conversationShowActions')?.checked
-  let list = conversationEntries
-  if (!showActions) list = list.filter(e => e.kind === 'in' || e.kind === 'out')
-  if (q) list = list.filter(e => (e.text || '').toLowerCase().includes(q))
-  // "Korábbiak betöltése" sits at the top so the operator can page further back;
-  // shown whenever the server still has older entries beyond the loaded window.
-  const olderBtn = conversationHasOlder
-    ? `<button id="conversationLoadOlder" class="conv-load-older">${t('conversation.load_more')}</button>`
-    : ''
-  if (!list.length) {
-    container.innerHTML = olderBtn || `<div class="conversation-empty">${t('conversation.empty')}</div>`
-  } else {
-    container.innerHTML = olderBtn + list.map(renderConvEntry).join('')
-  }
-  document.getElementById('conversationLoadOlder')?.addEventListener('click', loadOlderConversation)
-  if (opts.preserveScroll) {
-    // After prepending older messages, keep the previously-visible ones in place.
-    container.scrollTop = prevTop + (container.scrollHeight - prevH)
-  } else {
-    container.scrollTop = container.scrollHeight
-  }
-}
-
-function renderConvEntry(e) {
-  const ts = fmtConvTs(e.ts)
-  const txt = escapeHtml(e.text || '').replace(/\n/g, '<br>')
-  if (e.kind === 'in') {
-    return `<div class="conv-row conv-in"><div class="conv-bubble"><div class="conv-meta">Telegram be · ${ts}</div><div class="conv-text">${txt}</div></div></div>`
-  }
-  if (e.kind === 'out') {
-    const lbl = escapeHtml(e.label || t('messages.conv.reply_label'))
-    return `<div class="conv-row conv-out"><div class="conv-bubble"><div class="conv-meta">${lbl} · ${ts}</div><div class="conv-text">${txt}</div></div></div>`
-  }
-  if (e.kind === 'note') {
-    return `<div class="conv-row conv-note"><div class="conv-note-text">📝 ${txt}</div></div>`
-  }
-  return `<div class="conv-row conv-action"><div class="conv-action-text">⚙ ${txt}<span class="conv-action-ts">${ts}</span></div></div>`
-}
-
-document.getElementById('conversationClose')?.addEventListener('click', () => {
-  const overlay = document.getElementById('conversationOverlay')
-  if (overlay) closeModal(overlay)
-})
-document.getElementById('conversationSearch')?.addEventListener('input', () => renderConversation())
-document.getElementById('conversationShowActions')?.addEventListener('change', () => renderConversation())
-document.getElementById('conversationRefresh')?.addEventListener('click', () => loadConversation())
 
 // loadFederationPage, initFederation imported from ./modules/federation.js (S-13b)
 // ── Page registration + boot ──────────────────────────────────────────────────
